+++ llarp/handlers/exit.cpp
@@ -487,7 +487,7 @@ namespace llarp
         pkt.UpdateIPv6Address(from, m_IfAddr);
       else
         pkt.UpdateIPv4Address(
-            xhtonl(net::IPPacket::TruncateV6(from)), xhtonl(net::IPPacket::TruncateV6(m_IfAddr)));
+            xhtonl(net::TruncateV6(from)), xhtonl(net::TruncateV6(m_IfAddr)));
       return llarp_ev_tun_async_write(&m_Tun, pkt.Buffer());
     }
